+++ terraform/environments/examples/s3/versions.tf
@@ -1,0 +1,12 @@
+terraform {
+  required_version          = "~> 1.0.5"
+  backend "s3" {
+    key                   = "s3.tfstate"
+  }
+  required_providers {
+    aws                   = {
+      source              = "hashicorp/aws"
+      version             = "~> 3.57"
+    }
+  }
+}