+++ terraform/modules/ec2_asg/variables.tf
@@ -1,0 +1,14 @@
+# TFVars
+variable tf_vars {}
+
+# Roles
+variable ami_id {}
+variable base_aws_tags {}
+variable iam_instance_profile {}
+variable key_pair {}
+variable name {type = string}
+variable security_group_ids {}
+variable sns_topic {}
+variable subnet_ids {}
+variable user_data {}
+variable vpc {}