+++ terraform/modules/iam_user/local.tf
@@ -1,0 +1,9 @@
+locals {
+  aws_region                  = data.aws_region.current.name
+  aws_tags                    = merge(
+    var.base_aws_tags,
+    {
+      Terraform_Module          = "iam_user"
+    }
+  )
+}