+++ terraform/modules/route53_record/variables.tf
@@ -1,0 +1,6 @@
+# Role
+variable name {}
+variable records {}
+variable ttl {}
+variable type {}
+variable zone_id {}