+++ terraform/modules/ssm_parameter/outputs.tf
@@ -1,0 +1,4 @@
+output "ssm_parameter" {
+  value       = aws_ssm_parameter.default
+  // object
+}