+++ terraform/modules/vpn_endpoint/outputs.tf
@@ -1,0 +1,4 @@
+output vpn_endpoint {
+  value       = aws_ec2_client_vpn_endpoint.default
+  // object
+}